added plots for energy flux and co2 predictions

+++ exp.py
@@ -24,15 +24,16 @@ def filter_df_by_years(df, years):
 ###############################################
 
 REL_PATH = os.path.dirname(os.path.realpath(__file__)) + "/"
-CIMIS_DATAFILE = REL_PATH + "input/cimis.csv"
+CIMIS_DATAFILE = REL_PATH + "data/CIMIS/20160413-CIMIS-TWT-daily-2001-2016.csv"
 WESTPOND_DATAFILE = REL_PATH + "input/WP_2012195to2015126_L3.mat"
-WESTPOND_REFL_VEG_IDX = REL_PATH + "input/MODIS/2016-03-25-US-Tw1-MOD13Q1.csv"
-WESTPOND_LST_EM = REL_PATH + "input/MODIS/2016-03-25-US-Tw1-MOD11A2.csv"
+WESTPOND_REFL_VEG_IDX = REL_PATH + "data/MODIS/2016-04-13-US-Tw1-MOD13Q1.csv"
+WESTPOND_LST_EM = REL_PATH + "data/MODIS/2016-04-13-US-Tw1-MOD11A2.csv"
 MAYBERRY_DATAFILE = REL_PATH + "input/MB_2010287to2016055_L3.mat"
-MAYBERRY_REFL_VEG_IDX = REL_PATH + "input/MODIS/2016-03-25-US-Myb-MOD13Q1.csv"
-MAYBERRY_LST_EM = REL_PATH + "input/MODIS/2016-03-25-US-Myb-MOD11A2.csv"
+MAYBERRY_REFL_VEG_IDX = REL_PATH + "data/MODIS/2016-04-13-US-Myb-MOD13Q1.csv"
+MAYBERRY_LST_EM = REL_PATH + "data/MODIS/2016-04-13-US-Myb-MOD11A2.csv"
+TOWER_LWI = REL_PATH + "data/Tower-data/2016-04-14_daily_TOWER_LW.csv"
 
-def get_exp1_data(save_filename="input/exp1.csv"):
+def get_exp1_data(save_filename="data/exp1.csv"):
     if os.path.isfile(REL_PATH + save_filename):
         return pd.read_csv(REL_PATH + save_filename, index_col=0)
     cimis_df = utils.process_cimis(CIMIS_DATAFILE)
@@ -52,5 +53,23 @@ def get_exp1_data(save_filename="input/exp1.csv"):
                             mb_rf_vgi_df,
                             wp_lst_em_df,
                             mb_lst_em_df])
+    df.to_csv(REL_PATH + save_filename)
+    return df
+
+def get_wp_energy_data(save_filename="data/wp_energy.csv"):
+    if os.path.isfile(REL_PATH + save_filename):
+        return pd.read_csv(REL_PATH + save_filename, index_col=0)
+    cimis_df = utils.process_cimis(CIMIS_DATAFILE)
+
+    wp_df = utils.process_tower(WESTPOND_DATAFILE, prefix="wp")
+    wp_rf_vgi_df = utils.process_modis_reflectance_veg_index(WESTPOND_REFL_VEG_IDX, prefix="wp")
+    wp_lst_em_df = utils.process_modis_lst_emissivity(WESTPOND_LST_EM, prefix="wp")
+
+    tower_lwi = utils.process_tower_lwi(TOWER_LWI, interpolate_missing=True)
+    df = utils.merge_dataframes([cimis_df,
+                                wp_df,
+                                wp_rf_vgi_df,
+                                wp_lst_em_df,
+                                tower_lwi])
     df.to_csv(REL_PATH + save_filename)
     return df
